+++ src/lab1/universe_generation.cpp
@@ -41,9 +41,9 @@ namespace cadlabs {
 
     void original(
             int num_particles,
-            float *mass,
-            float *x_pos, float *y_pos,
-            float *x_vel, float *y_vel) {
+            double *mass,
+            double *x_pos, double *y_pos,
+            double *x_vel, double *y_vel) {
 
         double total_particle = num_particles;
         for (int i = 0; i < num_particles; i++) {
@@ -53,12 +53,12 @@ namespace cadlabs {
             x_vel[i] = y_pos[i];
             y_vel[i] = x_pos[i];
 #else
-            x_pos[i] = (float)i * (float)(2.0 / num_particles) - 1.0f;
+            x_pos[i] = i * (2.0 / num_particles) - 1.0f;
             y_pos[i] = 0.0;
             x_vel[i] = 0.0;
             y_vel[i] = x_pos[i];
 #endif
-            mass[i] = (float)(1.0 + (num_particles + i) / total_particle);
+            mass[i] = 1.0 + (num_particles + i) / total_particle;
         }
     }
 
@@ -90,19 +90,19 @@ namespace cadlabs {
 
     void sphere(
             int num_particles,
-            float *mass,
-            float *x_pos, float *y_pos) {
+            double *mass,
+            double *x_pos, double *y_pos) {
 
         int random = (rand() * num_particles);
-        float o = 2 / (double) num_particles;
-        float increment = PI * (3.0 - sqrt(5));
+        double o = 2 / (double) num_particles;
+        double increment = PI * (3.0 - sqrt(5));
 
         int rbase = MAX(1, MIN(num_particles / 1000, 10));
 
         for (int i = 0; i < num_particles; i++) {
-            float y = (((float)i * o) - 1) + (o / 2);
-            float r = sqrt( (float)rbase - pow(y, 2.0f));
-            float phi = (float)((i + random) % num_particles) * increment;
+            double y = ((i * o) - 1) + (o / 2);
+            double r = sqrt( rbase - pow(y, 2.0f));
+            double phi = ((i + random) % num_particles) * increment;
 
             x_pos[i] = cos(phi) * r;
             y_pos[i] = sin(phi) * r;
@@ -127,7 +127,7 @@ namespace cadlabs {
         for (int i = 0; i < num_particles; i++) {
             particle_t *particle = &particles[i];
 
-            double r = (float) ((rand() / RAND_MAX * radius) + RADIUS_OFFSET);
+            double r = rand() / RAND_MAX * radius + RADIUS_OFFSET;
 
             double alpha = rand() * 2 * PI;
 
@@ -150,9 +150,9 @@ namespace cadlabs {
 
     void rotating_disc(
             int num_particles,
-            float *mass,
-            float *x_pos, float *y_pos,
-            float *x_vel, float *y_vel) {
+            double *mass,
+            double *x_pos, double *y_pos,
+            double *x_vel, double *y_vel) {
 
         static const float RADIUS_OFFSET = 0.05f;
         static const float velocityMultiplier = 1.3;
@@ -164,7 +164,7 @@ namespace cadlabs {
         x_vel[0] = 1;
 
         for (int i = 0; i < num_particles; i++) {
-            double r = (float) ((rand() / RAND_MAX * radius) + RADIUS_OFFSET);
+            double r = rand() / RAND_MAX * radius + RADIUS_OFFSET;
             double alpha = rand() * 2 * PI;
 
             x_pos[i] = (float) (cos(alpha) * r);
@@ -172,7 +172,7 @@ namespace cadlabs {
             mass[i] = 1.0f + (num_particles + i) / num_particles;
 
             // orbital velocity
-            float v0 = (float) sqrt((1 + mass[i]) / (r * r * r)) * velocityMultiplier;
+            double v0 = sqrt((1 + mass[i]) / (r * r * r)) * velocityMultiplier;
 
             // rotate by 90°
             x_vel[i] = y_pos[i] * v0;
